hide internal parts of Input into InternalInput
